agents/systemcraft/backend/requirements.txt
# ════════════════════════════════════════════════
# 智码学长 SystemCraft - Python 依赖清单
# ════════════════════════════════════════════════
# 安装方式: pip install -r requirements.txt

# === LangGraph 核心 ===
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# === Web 框架 ===
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0

# === 工具库 ===
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0             # 终端美化输出，调试用
